[#6] Refactored player turning into using transform/rotation + some minor cleanup.
