+++ src/engine/debug.rs
@@ -1,8 +1,11 @@
 use std::collections::HashMap;
-use crate::engine::common::BoardPosition;
-use crate::simulation::players::Player;
+use crate::engine::common::*;
+use crate::simulation::players::{Player, FacingDirection};
 
 use bevy::prelude::*;
+use bevy::sprite::Mesh2dHandle;
+
+use crate::engine::config::default_entity_size;
 
 /// Debug system for logging if there are multiple players occupying the same tile
 fn log_positioning_conflicts(player_query: Query<&BoardPosition, With<Player>>) {
@@ -15,5 +18,34 @@ fn log_positioning_conflicts(player_query: Query<&BoardPosition, With<Player>>) 
             conflict_map.insert(*player_pos, num_players);
             error!(name: "kek", "Position {:#?} contains {} players!!", player_pos, num_players);
         }
+    }
+}
+
+// A function for producing a triangle mesh facing one of four cardinal directions
+pub fn triangle_facing(
+    direction: &FacingDirection,
+    meshes: &mut ResMut<Assets<Mesh>>,
+) -> Mesh2dHandle {
+    match *direction {
+        FacingDirection::Up => Mesh2dHandle(meshes.add(Triangle2d::new(
+            Vec2::new(0., default_entity_size() / 2.),
+            Vec2::new(-default_entity_size() / 2., -default_entity_size() / 2.),
+            Vec2::new(default_entity_size() / 2., -default_entity_size() / 2.),
+        ))),
+        FacingDirection::Right => Mesh2dHandle(meshes.add(Triangle2d::new(
+            Vec2::new(-default_entity_size() / 2., default_entity_size() / 2.),
+            Vec2::new(default_entity_size() / 2., 0.),
+            Vec2::new(-default_entity_size() / 2., -default_entity_size() / 2.),
+        ))),
+        FacingDirection::Down => Mesh2dHandle(meshes.add(Triangle2d::new(
+            Vec2::new(-default_entity_size() / 2., default_entity_size() / 2.),
+            Vec2::new(default_entity_size() / 2., default_entity_size() / 2.),
+            Vec2::new(0., -default_entity_size() / 2.),
+        ))),
+        FacingDirection::Left => Mesh2dHandle(meshes.add(Triangle2d::new(
+            Vec2::new(-default_entity_size() / 2., 0.),
+            Vec2::new(default_entity_size() / 2., default_entity_size() / 2.),
+            Vec2::new(default_entity_size() / 2., -default_entity_size() / 2.),
+        ))),
     }
 }
